% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcolbyname.R
\name{addcolbyname}
\alias{addcolbyname}
\title{Add an empty column to a dataframe}
\usage{
addcolbyname(data, name)
}
\arguments{
\item{data}{Dataframe to which to append empty column}

\item{name}{Name of the new column}
}
\description{
This function will append an empty (NA-filled) column to the dataframe with name "name"
}
\examples{
addnumbersequence(data = data, name = 'New Column Name')
}
\keyword{addcolbyname}
