% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNDAhead.R
\name{addNDAhead}
\alias{addNDAhead}
\title{Add header to NDA upload files}
\usage{
addNDAhead(data)
}
\description{
This function adds the header to data to be uploaded to NDA
}
\examples{
addNDAhead()
}
\keyword{addNDAhead}
\keyword{ndatools}
