% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnumbersequence.R
\name{addnumbersequence}
\alias{addnumbersequence}
\title{Add header to NDA upload files}
\usage{
addnumbersequence(text, delimiter, range)
}
\arguments{
\item{test}{delimiter range}
}
\description{
This function will take text and add a range of numbers to it so the user can copy and paste into their script. For example if you have a 21-item questionnaire (e.g., pqb) and want to add up all of the items, you can specify the text (e.g.., "pqb"), the numbers that follow (e.g., 1:21), and the delimiter (e.g., '+'), and you'll get a nice list that you can print and copy and paste.
}
\examples{
addnumbersequence(text = 'pqb', delimiter = ' + ', range = c(1:21))
addnumbersequence()
}
\keyword{letters}
\keyword{numbers}
\keyword{sequence}
