% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addsamp.r
\name{addsamp}
\alias{addsamp}
\title{Add brainspan sample (donor) to mappable df.}
\usage{
addsamp(mappable)
}
\description{
This function adds sample (donor) to mappable df and is necessary for expression.compare function.
}
\examples{
addsamp()
}
\keyword{add}
\keyword{brainspan}
\keyword{donor}
\keyword{mappable}
