% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustmatrix.r
\name{adjustmatrix}
\alias{adjustmatrix}
\title{Make easy peas}
\usage{
adjustmatrix(x)
}
\description{
This function replaces p-values with numbers according to a threshold
}
\examples{
adjustmatrix()
}
\keyword{adjustmatrix}
