% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustmatrixmore.r
\name{adjustmatrixmore}
\alias{adjustmatrixmore}
\title{Make easy peas}
\usage{
adjustmatrixmore(x)
}
\description{
This function replaces p-values with numbers according to a threshold
}
\examples{
adjustmatrixmore()
}
\keyword{adjustmatrix}
