% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenupdatehgnclabels.r
\name{allenupdatehgnclabels}
\alias{allenupdatehgnclabels}
\title{Updates deprecated HGNC gene symbols in allen expression}
\usage{
allenupdatehgnclabels(x)
}
\description{
This function updates old HGNC gene symbols in allen expression (via freesurfer paper).
}
\examples{
allenupdatehgnclabels()
}
\keyword{allen}
\keyword{atlas,}
\keyword{hgnc}
\keyword{labels}
