% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2mappable.r
\name{convert2mappable}
\alias{convert2mappable}
\title{Convert gene expression data into mappable data..}
\usage{
convert2mappable(x, ageref, include.dupes = T)
}
\arguments{
\item{x}{ageref include.dupes}
}
\description{
This function converts expression data to interpretable/mappabale data
}
\examples{
convert2mappable()
}
\keyword{convert}
\keyword{expression}
\keyword{mappable}
