% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2num.r
\name{convert2num}
\alias{convert2num}
\title{Converts columns to numeric}
\usage{
convert2num(x, skipcols = NULL)
}
\arguments{
\item{x}{skipcols}
}
\description{
This function converts character/factor columns in a data frame (x) to class numeric. The function will assume the first column is the subject key and will convert all but the first column to numeric values. Inputting skipcols will tell the function which columns to skip i.e. non-numeric ones.
}
\examples{
convert2num()
}
\keyword{character,}
\keyword{conversion,}
\keyword{factor}
\keyword{from}
\keyword{numeric,}
