% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationcolormat.r
\name{correlationcolormat}
\alias{correlationcolormat}
\title{Assist heatmap function}
\usage{
correlationcolormat(ref, empty, slide = T, max = NULL)
}
\arguments{
\item{ref}{empty}
}
\description{
This function is for the heatmap function, helps to pick colors and breaks
}
\examples{
correlationcolormat()
}
\keyword{color,}
\keyword{heatmap}
