% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.r
\name{direction}
\alias{direction}
\title{Determine lme4 direction for heatmap}
\usage{
direction(x, y)
}
\arguments{
\item{x}{y}
}
\description{
This function determines direction of the effect from lme4 betas
}
\examples{
direction()
}
\keyword{directionofeffect}
