% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillinNAs.R
\name{fillinNAs}
\alias{fillinNAs}
\title{Fill NA values with specified values}
\usage{
fillinNAs(data, cols, fill = NULL)
}
\arguments{
\item{data}{cols}
}
\description{
This function is used in conjunction with pullNAaction and will replace NAs with NDAR-specified missing values
}
\examples{
fillinNAs()
}
\keyword{fillinNAs}
\keyword{ndatools}
