% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.r
\name{get_legend}
\alias{get_legend}
\title{Grab legend in glob format from ggplot plot}
\usage{
get_legend(myggplot)
}
\arguments{
\item{myggplot}{a ggplot object from which to grab legend}
}
\description{
This function grabs a legend from your ggplot plot to be used later with gridExtra functions. This was taken from http://stackoverflow.com/questions/12539348/ggplot-separate-legend-and-plot. makePie() calls this function, so they must be loaded together. I'm not sure how exactly this function can be helpful outside of the context of makePie().
}
\keyword{get_legend}
\keyword{makePie}
