% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getridofnas.r
\name{getridofnas}
\alias{getridofnas}
\title{Replaces 'fake' NA values with 'true' NAs}
\usage{
getridofnas(df, nas = NA)
}
\arguments{
\item{df}{nas}
}
\description{
Sometimes NAs are coded as 999 or 777 or just as blank spaces. This function replaces those substitute or 'fake' values with 'true' NAs that R can interpret as such. the nas arugment by default is set to a 'true' NA. You can set it to a list. I usually do nas=c(777,999,''). Note that there is no space between the two single quotes.
}
\examples{
getridofnas(df = data, nas = c(777,999,''))
getridofnas()
}
\keyword{nas}
\keyword{replace}
