% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.oddsratio.r
\name{glm.oddsratio}
\alias{glm.oddsratio}
\title{Add ORs and 95\% CIs to glm model summary}
\usage{
glm.oddsratio(mod)
}
\description{
This function takes a glm model and adds odds ratio and 95\% CIs to model output
}
\examples{
glm.oddsratio()
}
\keyword{glm,}
\keyword{oddsratio,}
\keyword{summary}
