% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtLmerInterpret.r
\name{gtLmerInterpret}
\alias{gtLmerInterpret}
\title{Function to print a pretty version of your model output}
\usage{
gtLmerInterpret(
  model,
  ci = T,
  var = NULL,
  or.only = F,
  r2 = F,
  alternate.ci = F,
  row.names = NULL,
  rowname.dict = NULL,
  return.table = F
)
}
\arguments{
\item{model}{This is a model of class lm or lmerMod}

\item{ci}{If True, function will print confidence intervals around the estimates. Defaults to True. Can increase runtime semi-substantially}

\item{var}{Specify a variable name or vector of variable names for which to calculate confidence intervals. This can make things more efficient if you only need the CI of your main variable. Can be useful to run summary(model) first to see exactly what the variable name looks like in the coefficient table. For example, sometimes it will be "variableNameLevelName" i.e., variable name followed by level name if its a categorical variable.}

\item{or.only}{If true and the input model is of the class glmerMod, then only the odds ratios will be reported (i.e., not the associated coefficient estimate aka log odds)}

\item{r2}{If true, model r-squared will be returned}

\item{row.names}{Can set this to a list of custom row names for prettier gt tables}

\item{rowname.dict}{Can provide a dictionary (in list form) that has names as the 'ugly'/messy names that appear in the gt output and the values as prettier versions.}

\item{return.table}{if True, then return an object that stores the table version (e.g., for csvs)}

\item{alternate.if}{If true, confidence interval will be calculated using the formula: estimate ± 1.96 * standard error}
}
\description{
This function prints the output of linear models in a pretty fashion using gt
}
