% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtLmInterpret.r
\name{gtLmInterpret}
\alias{gtLmInterpret}
\title{Make a pretty gt table out of lm (or glm or lmerTest) model output}
\usage{
gtLmInterpret(model)
}
\arguments{
\item{model}{An object (model) of class lm or lmerModLmerTest}
}
\description{
This function takes a model as input and outputs the coefficients from the model in a nice table
}
\keyword{gt}
\keyword{kableLmerInterpret}
\keyword{summary}
\keyword{table}
