% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_much_longer.r
\name{how_much_longer}
\alias{how_much_longer}
\title{Estimate the amount of time left in a loop}
\usage{
how_much_longer(toti)
}
\arguments{
\item{toti}{This represents the length of the vector the loop iterates over. So if the loop will iterate over a vector of length 10 (i.e., 10 iterations), toti would equal 10.}
}
\description{
This function is embedded in a loop and will print updates on the estimated time remaining in the loop to the R console.
}
\details{
This function can be embedded in any loop to estimate how much time is left before the loop finishes. See examples for how to use within a loop. It estimates the remaining time by creating and continually updating a vector containing the amount of time each loop takes and then taking the mean of the vector and multiplying by the number of iterations remaining. Probs one of the most powerful funcs I've written hehe
}
\examples{
dvs = sample(LETTERS, 10)

timer <- how_much_longer(toti = length(dvs))

for (i in seq_along(dvs)) {
  dv = dvs[i]
  time = system.time({
    final_df[,i] = some_function(dv = dv)
  })
  timer(time, i)
}

}
\author{
Original code by Dylan E. Hughes
}
