% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeheatmap.r
\name{lmeheatmap}
\alias{lmeheatmap}
\title{Make lme4 heatmap}
\usage{
lmeheatmap(dvs, ivs, pcs, age, sex, x)
}
\arguments{
\item{dvs}{ivs pcs age sex x}
}
\description{
This function makes a heatmap (cbcl vs prs) with scaled principal components, age, sex, and site as a random effect
}
\examples{
lmeheatmap()
}
\keyword{prs}
