% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.interpret.ci.r
\name{lmer.interpret.ci}
\alias{lmer.interpret.ci}
\title{Add 95\% CIs to lmer.interpret output}
\usage{
lmer.interpret.ci(mod)
}
\description{
This function takes a lmer mod and appends 95\% CIs to lmer.interpret output
}
\examples{
lmer.interpret.ci()
}
\keyword{confidence}
\keyword{intervals}
\keyword{lmer,}
