% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_interpret.r
\name{lmer.interpret}
\alias{lmer.interpret}
\title{Interpret lmer}
\usage{
lmer.interpret(lmermodel)
}
\arguments{
\item{lmermodel}{model}
}
\description{
This function calculates the p-values of a lmer model from the t-values of the fixed effects coefficients
}
\examples{
lmer.interpret()
}
\keyword{lmermodel}
