% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_vianova.r
\name{lmer.vianova}
\alias{lmer.vianova}
\title{Interpret lmer with car::Anova func}
\usage{
lmer.vianova(lmermodel)
}
\arguments{
\item{lmermodel}{model}
}
\description{
This function calculates the p-values of a lmer model from the t-values of the fixed effects coefficients
}
\examples{
lmer.vianova()
}
\keyword{lmermodel}
