% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeroddsratio.r
\name{lmeroddsratio}
\alias{lmeroddsratio}
\title{Generate odds ratio outputs from glmer model}
\usage{
lmeroddsratio(mod)
}
\description{
This function takes a glmer model and outputs odds ratios and CIs
}
\examples{
lmeroddsratio()
}
\keyword{OR}
