% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopModThroughVars.r
\name{loopModThroughVars}
\alias{loopModThroughVars}
\title{Function to loop through DVs and IVs and return results from lme4::lmer() models}
\usage{
loopModThroughVars(dvs, ivs, rownames, colnames, covariates, data, na.omit = F)
}
\arguments{
\item{dvs}{A vector of characters corresponding to variables in your dataframe that represent dependent variables}

\item{ivs}{A vector of characters corresponding to your main predictors/independent variables}

\item{rownames}{Vector of characters specifying row names of the resultant table. These will correspond to your main predictors/IVs}

\item{colnames}{Vector of characters specifying column names of the resultant table. These correspond to your outcomes/DVs}

\item{data}{Dataframe in which your variables live}

\item{na.omit}{If true, will first remove instances of NAs using the na.omit function. Defaults to FALSE.}
}
\description{
This function iterates through mixed effects models with a specified set of dependent and independent variables as well as covariates.
}
\examples{
master <- read.csv('path/to/master/file')
all.dvs <- c('flanker','processingSpeed','dimChangeCard', 'DistressScoreSum')
cpbrain.pgs <- colnames(select(master, contains('z.CP_M')))
row.names=substr(cpbrain.pgs, start = 6, stop = nchar(cpbrain.pgs)) # picks the 6th through last character of each element of cpbrain.pgs
col.names = substr(all.dvs, start = 1, stop = 15) # picks 1st through 15th character of each element of all.dvs
covariates = c('z.pc1','z.pc2','z.pc3','z.pc4','z.pc5','interview_age_new','demo_sex_v2','(1|site_id_l/rel_family_id/subjectkey)')
cpbrainspan <- loopModThroughVars(dvs = all.dvs, ivs = cpbrain.pgs, rownames=row.names, colnames = col.names, covariates = covariates, data = master)
}
\keyword{loopMod}
\keyword{loopThroughMod}
