% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopPieThroughYears.r
\name{loopPieThroughYears}
\alias{loopPieThroughYears}
\title{Make pie charts for multiple timepoints at once}
\usage{
loopPieThroughYears(data, times)
}
\arguments{
\item{data}{A vector within a dataframe to be pie'd}

\item{times}{A vector containing the timepoints you would like to iterate through}
}
\description{
This function iterates through user-provided timepoints and generates a pie chart for each timepoint. Assumes your data is in long format. Calls makePie() argument. On the last iteration, the loop will automatically grab the legend from the plot. The output is a bit wonky. See "Examples" for an example of how to use it with grid.arrange().
}
\examples{
events = levels(as.factor(dataframe$time_variable))[2:5] # Will grab the second through fifth timepoints in your data assuming your time variable is called "time_variable" within your dataframe.
pieList <- loopPieThroughYears(data = dataframe$categorical_variable1, times = events)
grid.arrange(pieList[[1]], pieList[[2]], pieList[[3]], pieList[[4]][1]$plot, pieList[[4]][2]$legend, nrow = 3, ncol = 2, layout_matrix = rbind(c(1,2), c(3,4), c(5,5)))
}
\keyword{loop}
\keyword{loopPieThroughYears}
\keyword{makePie}
\keyword{pie}
