% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIntoTable.r
\name{makeIntoTable}
\alias{makeIntoTable}
\title{Make table output from loopModThroughVars() function into a pretty table}
\usage{
makeIntoTable(obj = NULL, betas = NULL, peas = NULL, fdr = NULL)
}
\arguments{
\item{obj}{An object returned from loopModThroughVars() containing beta values, raw p-vals, and corrected p-vals. Defaults to NULL so you have the choice of providing your own values (i.e., not generated via the loopMod.. func)}

\item{betas}{Matrix of betas. Defaults to NULL. If you're using this function to process output from loopModThroughVars(), you can ignore this argument as well as peas and fdr arguments.}

\item{peas}{Matrix of raw p-values. Defaults to NULL. See note above (under betas argument description).}

\item{fdr}{Matrix of corrected p-values. Defaults to NULL. See note above (under betas argument description).}
}
\description{
This function takes output from dylanfuncs::loopModThroughVars() and makes it into a pretty table. not much else. You can also manually provide a table of betas, p-vals, and corrected p-vals with the same dimensions.
}
\examples{
master <- read.csv('path/to/master/file')
all.dvs <- c('flanker','processingSpeed','dimChangeCard', 'DistressScoreSum')
cpbrain.pgs <- colnames(select(master, contains('z.CP_M')))
row.names=substr(cpbrain.pgs, start = 6, stop = nchar(cpbrain.pgs)) # picks the 6th through last character of each element of cpbrain.pgs
col.names = substr(all.dvs, start = 1, stop = 15) # picks 1st through 15th character of each element of all.dvs
covariates = c('z.pc1','z.pc2','z.pc3','z.pc4','z.pc5','interview_age_new','demo_sex_v2','(1|site_id_l/rel_family_id/subjectkey)')
cpbrainspan <- loopModThroughVars(dvs = all.dvs, ivs = cpbrain.pgs, rownames=row.names, colnames = col.names, covariates = covariates, data = master)
makeIntoTable(obj = cpbrainspan)
}
\keyword{loopMod}
\keyword{loopThroughMod}
\keyword{makeIntoTable}
\keyword{makePretty}
