% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediDiagram.r
\name{makeMediDiagram}
\alias{makeMediDiagram}
\title{Makes mediation diagram}
\usage{
makeMediDiagram(medvals, a.lab, b.lab, c.lab)
}
\arguments{
\item{medvals}{this is the output from pull_important_medvals}

\item{a.lab}{the label to be added to the "a" box which represents the main predictor}

\item{b.lab}{the label to be added to the "b" box which represents the mediator variable}

\item{c.lab}{the label top be added to the "c" box which represents the main outcome}
}
\description{
Takes output from pull_important_medvals and plots them in a digestible mediation path diagram. Mediation models must be from the mediation package
}
\keyword{makeMediDiagram,}
\keyword{mediation,}
\keyword{pull_important_medvals}
