% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdupegenes.r
\name{markdupegenes}
\alias{markdupegenes}
\title{Marks duplicate genes}
\usage{
markdupegenes(x)
}
\description{
This function marks duplicate genes and is necessary for convert2mappable function.
}
\examples{
markdupegenes()
}
\keyword{convert2mappable}
\keyword{dupes}
\keyword{genes}
