% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninestonas.r
\name{ninestonas}
\alias{ninestonas}
\title{Turn 999s and 777s to NAs}
\usage{
ninestonas(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function turns 999s and 777s into Nas
}
\examples{
ninestonas()
}
\keyword{nas}
