% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickcolors.fhm.r
\name{pickcolors.fhm}
\alias{pickcolors.fhm}
\title{Make colors for heatmap}
\usage{
pickcolors.fhm(x)
}
\description{
This function selects colors for heatmap
}
\examples{
pickcolors.fhm()
}
\keyword{colors}
