% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotprsdist.r
\name{plotprsdist}
\alias{plotprsdist}
\title{Plot PRS distributions across thresholds}
\usage{
plotprsdist(df, prsname)
}
\arguments{
\item{df}{prsname}
}
\description{
This function reads in plink polygenic risk scores
}
\examples{
plotprsdist()
}
\keyword{distribution}
