% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullNAaction.R
\name{pullNAaction}
\alias{pullNAaction}
\title{Pull NDA-specified NA value}
\usage{
pullNAaction(elementname)
}
\description{
This function looks in NDA def files to figure out what value NDA wants missing data to be coded as
}
\examples{
pullNAaction()
}
\keyword{ndatools}
\keyword{pullNAaction}
