% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullorsgen.r
\name{pullorsgen}
\alias{pullorsgen}
\title{Pull ORs from glmer mice models}
\usage{
pullorsgen(x)
}
\arguments{
\item{x}{model}
}
\description{
This function takes a glmer mice model as an argument and prints ORs and Peas
}
\examples{
pullorsgen()
}
\keyword{glmer,}
\keyword{mice,}
\keyword{odds}
\keyword{ors,}
\keyword{ratios}
