% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullOutcome.r
\name{pullOutcome}
\alias{pullOutcome}
\title{Pull outcome variable name from input model}
\usage{
pullOutcome(model)
}
\arguments{
\item{model}{An object (model) of class lm or lmerModLmerTest or glm}
}
\description{
This function takes a model (lm, lmerMod, or glm) as input and outputs the name of the outcome variable
}
\keyword{gt}
\keyword{kableLmerInterpret}
\keyword{pullOutcome}
\keyword{summary}
\keyword{table}
