% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpcas.r
\name{readpcas}
\alias{readpcas}
\title{Read PRS ranges}
\usage{
readpcas(x, nofsubjects)
}
\arguments{
\item{x}{nofsubjects}
}
\description{
This function reads in plink polygenic risk scores
}
\examples{
readpcas()
}
\keyword{prs}
