% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.nda.R
\name{recode.nda}
\alias{recode.nda}
\title{Recode 22q data to match NDA data structures}
\usage{
\method{recode}{nda}(raw, x, y)
}
\arguments{
\item{raw}{x y}
}
\description{
This function recodes 22q data to make it compatible with NDA submission
}
\examples{
recode.nda(raw = raw, x = 22q.coding, y = nda.coding)
recode.nda()
}
\keyword{recode.nda}
