% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamecbclcolnames.r
\name{renamecbclcolnames}
\alias{renamecbclcolnames}
\title{Update CBCL names to the Roffman convention}
\usage{
renamecbclcolnames(df, timepoint)
}
\arguments{
\item{df, }{cols, timepoint}
}
\description{
This function takes the ABCD cbcl names and turns them into our convention (i.e. capital, preceded by timepoint ex: BL, Y2, etc)
}
\examples{
renamecbclcolnames()
}
\keyword{cbcl,}
\keyword{column}
\keyword{names}
\keyword{renaming,}
