% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameprsice.r
\name{renameprsice}
\alias{renameprsice}
\title{Rename column names by prs}
\usage{
renameprsice(df, prs)
}
\arguments{
\item{dataframe}{prsname}
}
\description{
This function attaches the prs name to column names of a dataframe
}
\examples{
renameprsice()
}
\keyword{prs}
\keyword{prsice}
