% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversescore.r
\name{reversescore}
\alias{reversescore}
\title{Reverse scores items}
\usage{
reversescore(x, levels)
}
\arguments{
\item{x}{levels}
}
\description{
This function reverse scores items. The levels argument should be the levels of the original variables e.g. if you have a question where 1 = yes, 2 = maybe, 3 = no, and you want to flip those around, the levels argument would be c(1,2,3).
}
\examples{
reversescore()
}
\keyword{reverse}
\keyword{score}
