% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughMediation.r
\name{roughMediation}
\alias{roughMediation}
\title{Performs a rough mediation using simple linear regression framework}
\usage{
roughMediation(data, x, y, mediator, covariates, treat = NULL)
}
\arguments{
\item{data}{Dataframe containing variables of interest}

\item{x}{Your x variable, aka independent variable, as a character}

\item{y}{Your y variable, aka dependent variable, as a character}

\item{mediator}{Your mediator variable as a character}

\item{covariates}{Your covariates}

\item{treat}{If your x variable is categorical, this specifies what the level we are interested in is. For example, if we have a gender variable with 3 levels, man, woman, and trans, and we are interested in the effect associated with the 'trans' level, we would provide the name of the level i.e., 'trans' to this argument}
}
\description{
Given x, a continuous y, a mediator, and optional covariates, this function will generate the total, direct, and indirect effect as well as provide the proportion of the total effect mediated by the mediator and the p-value of X given M. It will not provide the p-value associated with the indirect effect.
}
\keyword{mediation,}
\keyword{rough}
