% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsreformat.r
\name{rsreformat}
\alias{rsreformat}
\title{Convert snp names to rs format}
\usage{
rsreformat(sumstats, output)
}
\arguments{
\item{sumstats, }{output}
}
\description{
This function converts chr bp to rs
}
\examples{
rsreformat()
}
\keyword{a1a2}
\keyword{rs,}
\keyword{sap,}
