% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.letters.R
\name{select.letters}
\alias{select.letters}
\title{Select only letters from a string}
\usage{
\method{select}{letters}(x)
}
\description{
This function selects only letters from a string (i.e., removes numbers)
}
\examples{
select.letters()
}
\keyword{select.letters}
