% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripab.r
\name{stripab}
\alias{stripab}
\title{Strip AB bit from SUBJECTKEY}
\usage{
stripab(vec)
}
\description{
This function strips the "AB" bit from subject IDs output from PRSice
}
\examples{
stripab()
}
\keyword{ab}
\keyword{prsice}
\keyword{strip}
