% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectkey2nsubformat.r
\name{subjectkey2nsubformat}
\alias{subjectkey2nsubformat}
\title{Convert to nsub format}
\usage{
subjectkey2nsubformat(og.names)
}
\description{
This function takes converts the NDAR_INV subject key format to nsub-NDARINV subject key format
}
\examples{
subjectkey2nsubformat()
}
\keyword{key}
\keyword{subject}
