% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatefdr.r
\name{updatefdr}
\alias{updatefdr}
\title{Make fdr labels}
\usage{
updatefdr(x)
}
\description{
This function adjusts labels for heatmap to show fdr significant boxes
}
\examples{
updatefdr()
}
\keyword{fdrcorrect}
